chore(aio): move QUIET_REPORTER into the braintrust eval engine
QUIET_REPORTER (and its report_eval/report_run callbacks) is a braintrust-engine invariant, not a harness-reporting concern.
Relocate it from harness/reporting.py to engines/braintrust.py, where the other EvalAsync knobs already live, so reporting.py sheds one more braintrust import on its way to being braintrust-free.
